Move coordinator in _init_.py
